Consolidate login and registration to one form
